premier essai du 24 janvier

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: claville <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/20 11:28:59 by claville          #+#    #+#             */
-/*   Updated: 2017/11/27 16:39:41 by claville         ###   ########.fr       */
+/*   Updated: 2018/01/23 11:49:05 by claville         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *s)
+size_t		ft_strlen(const char *s)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s[i] != '\0')
